CLN: GH11271 move _get_handle, UTF encoders to io.common
